third_party/blink/web_tests/fast/dom/NodeList/nodelist-reachable-expected.txt

PASS document.body.childNodes.customProperty is 1
PASS document.getElementsByClassName("class").customProperty is 2
PASS document.getElementsByName("name").customProperty is 3
PASS document.getElementsByTagName("body").customProperty is 4
PASS document.querySelector("form").elements["radios"].customProperty is 5
PASS document.querySelector("input").labels.customProperty is 6
PASS successfullyParsed is true

TEST COMPLETE

 
